Don't ever display status bar if presenting view controller didn't.
Without this fix, the status bar would briefly be presented (and then immediately dismissed) during the transition animation to display an image. More importantly, this would cause the y-position of the original image to be off by the height of the status bar during the transition animation, causing a distracting jump.